remove the `--check` option
Removing this option because it's actually completely useless.

Why? Because an `.md` file needs to be created regardless of whether the updated site will be published or not. `zola check` will check links, etc. to make sure that none are broken. If this only checks, but doesn't create an `.md` file, errors will be thrown even if the site is set up correctly.